+++ .github/workflows/ci.yml
@@ -32,6 +32,11 @@ jobs:
       - uses: actions/checkout@v4
       - name: Update Rust toolchain
         run: rustup update
+      - uses: astral-sh/setup-uv@v5
+        with:
+          python-version: 3.11
+      - name: Install dependencies
+        run: uv pip install -e .
       - name: Run tests
         run: |
           cargo test
@@ -97,7 +102,7 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.9", "3.13"]
-        folder: ["tests/integration", "tests/example"]
+        folder: ["tests/example"]
         os: [ubuntu-latest, macos-latest, windows-latest]
     steps:
       - uses: actions/checkout@v4
