+++ Cargo.toml
@@ -37,9 +37,10 @@ karva_project = { path = "crates/karva_project" }
 markdown = { version = "1.0.0" }
 notify = { version = "8.0.0" }
 pretty_assertions = { version = "1.4.1" }
-pyo3 = { version = "0.25.0", features = ["auto-initialize"] }
+pyo3 = { version = "0.25.1", features = ["auto-initialize"] }
 rand = { version = "0.9.1" }
 regex = "1"
+rstest = { version = "0.18" }
 ruff_python_ast = { git = "https://github.com/astral-sh/ruff/", branch = "main" }
 ruff_python_parser = { git = "https://github.com/astral-sh/ruff/", branch = "main" }
 ruff_python_trivia = { git = "https://github.com/astral-sh/ruff/", branch = "main" }
