+++ crates/karva/Cargo.toml
@@ -10,14 +10,15 @@ repository = { workspace = true }
 authors = { workspace = true }
 license = { workspace = true }
 
+[lib]
+name = "karva"
+crate-type = ["rlib", "cdylib"]
+
 [dependencies]
 karva_cli = { workspace = true }
 karva_core = { workspace = true }
 
 pyo3 = { workspace = true }
-
-[features]
-extension-module = ["pyo3/extension-module"]
 
 [lints]
 workspace = true
