+++ crates/karva_cli/tests/cli.rs
@@ -4,9 +4,41 @@ use std::{
 };
 
 use anyhow::Context;
-use insta::internals::SettingsBindDropGuard;
-use insta_cmd::{assert_cmd_snapshot, get_cargo_bin};
+use insta::{allow_duplicates, internals::SettingsBindDropGuard};
+use insta_cmd::assert_cmd_snapshot;
+use rstest::rstest;
 use tempfile::TempDir;
+
+/// Find the karva wheel in the target/wheels directory.
+/// Returns the path to the wheel file.
+fn find_karva_wheel() -> anyhow::Result<PathBuf> {
+    let karva_root = Path::new(env!("CARGO_MANIFEST_DIR"))
+        .parent()
+        .and_then(|p| p.parent())
+        .ok_or_else(|| anyhow::anyhow!("Could not determine KARVA_ROOT"))?
+        .to_path_buf();
+
+    let wheels_dir = karva_root.join("target").join("wheels");
+
+    let entries = std::fs::read_dir(&wheels_dir)
+        .with_context(|| format!("Could not read wheels directory: {}", wheels_dir.display()))?;
+
+    for entry in entries {
+        let entry = entry?;
+        let file_name = entry.file_name();
+        if let Some(name) = file_name.to_str() {
+            if name.starts_with("karva-")
+                && std::path::Path::new(name)
+                    .extension()
+                    .is_some_and(|ext| ext.eq_ignore_ascii_case("whl"))
+            {
+                return Ok(entry.path());
+            }
+        }
+    }
+
+    anyhow::bail!("Could not find karva wheel in target/wheels directory");
+}
 
 struct TestCase {
     _temp_dir: TempDir,
@@ -16,11 +48,8 @@ struct TestCase {
 
 impl TestCase {
     fn new() -> anyhow::Result<Self> {
-        let temp_dir = TempDir::new()?;
+        let temp_dir = TempDir::with_prefix("karva-test-env")?;
 
-        // Canonicalize the tempdir path because macos uses symlinks for tempdirs
-        // and that doesn't play well with our snapshot filtering.
-        // Simplify with dunce because otherwise we get UNC paths on Windows.
         let project_dir = dunce::simplified(
             &temp_dir
                 .path()
@@ -28,6 +57,52 @@ impl TestCase {
                 .context("Failed to canonicalize project path")?,
         )
         .to_path_buf();
+
+        let karva_wheel = find_karva_wheel()?;
+
+        // Set up a bare uv project and install pytest, mimicking the Python TestEnv
+        let commands = [
+            vec![
+                "uv",
+                "init",
+                "--bare",
+                "--directory",
+                project_dir.to_str().unwrap(),
+            ],
+            vec![
+                "uv",
+                "add",
+                "pytest",
+                "--directory",
+                project_dir.to_str().unwrap(),
+            ],
+            vec!["uv", "sync", "--directory", project_dir.to_str().unwrap()],
+            vec![
+                "uv",
+                "pip",
+                "install",
+                "--directory",
+                project_dir.to_str().unwrap(),
+                karva_wheel.to_str().unwrap(),
+            ],
+        ];
+
+        for command in &commands {
+            let output = Command::new(command[0])
+                .args(&command[1..])
+                .current_dir(&project_dir)
+                .output()
+                .with_context(|| format!("Failed to run command: {command:?}"))?;
+            if !output.status.success() {
+                eprintln!(
+                    "Command failed: {:?}\nstdout:\n{}\nstderr:\n{}",
+                    command,
+                    String::from_utf8_lossy(&output.stdout),
+                    String::from_utf8_lossy(&output.stderr)
+                );
+                anyhow::bail!("Command failed: {:?}", command);
+            }
+        }
 
         let mut settings = insta::Settings::clone_current();
         settings.add_filter(&tempdir_filter(&project_dir), "<temp_dir>/");
@@ -40,6 +115,14 @@ impl TestCase {
             _temp_dir: temp_dir,
             _settings_scope: settings_scope,
         })
+    }
+
+    fn karva_bin(&self) -> PathBuf {
+        let venv_bin =
+            self.project_dir
+                .join(".venv")
+                .join(if cfg!(windows) { "Scripts" } else { "bin" });
+        venv_bin.join(if cfg!(windows) { "karva.exe" } else { "karva" })
     }
 
     fn with_files<'a>(files: impl IntoIterator<Item = (&'a str, &'a str)>) -> anyhow::Result<Self> {
@@ -80,8 +163,14 @@ impl TestCase {
     }
 
     fn command(&self) -> Command {
-        let mut command = Command::new(get_cargo_bin("karva"));
+        let mut command = Command::new(self.karva_bin());
         command.current_dir(&self.project_dir).arg("test");
+        command
+    }
+
+    fn command_with_args(&self, args: &[&str]) -> Command {
+        let mut command = self.command();
+        command.args(args);
         command
     }
 }
@@ -684,6 +773,1893 @@ fn test_assertion_in_imported_function() -> anyhow::Result<()> {
      |   assert value > 0, f'Expected positive value, got {value}'
 
     Failed tests: 1
+
+    ----- stderr -----
+    "#);
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_fixture_initialization_order(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="session")
+                def session_calculator() -> Calculator:
+                    print("Session calculator initialized")
+                    return Calculator()
+
+                @fixture(scope="module")
+                def module_calculator() -> Calculator:
+                    print("Module calculator initialized")
+                    return Calculator()
+
+                @fixture(scope="package")
+                def package_calculator() -> Calculator:
+                    print("Package calculator initialized")
+                    return Calculator()
+
+                @fixture
+                def function_calculator() -> Calculator:
+                    print("Function calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_all_scopes(
+                session_calculator: Calculator,
+                module_calculator: Calculator,
+                package_calculator: Calculator,
+                function_calculator: Calculator,
+            ) -> None:
+                assert session_calculator.add(1, 2) == 3
+                assert module_calculator.add(1, 2) == 3
+                assert package_calculator.add(1, 2) == 3
+                assert function_calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Session calculator initialized
+    Package calculator initialized
+    Module calculator initialized
+    Function calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[test]
+fn test_empty_conftest() -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            r"
+            # Empty conftest file
+            "
+            .to_string(),
+        ),
+        (
+            "tests/conftest.py".to_string(),
+            r"
+            # Another empty conftest file
+            "
+            .to_string(),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_no_fixtures() -> None:
+                calculator = Calculator()
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+
+    ----- stderr -----
+    "#);
+
+    Ok(())
+}
+
+fn get_parametrize_function(package: &str) -> String {
+    if package == "pytest" {
+        "pytest.mark.parametrize".to_string()
+    } else {
+        "karva.tags.parametrize".to_string()
+    }
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_parametrize(#[case] package: &str) -> anyhow::Result<()> {
+    let case = TestCase::with_file(
+        "test_parametrize.py",
+        &format!(
+            r"
+        import {package}
+
+        @{parametrize_function}(('a', 'b', 'expected'), [
+            (1, 2, 3),
+            (2, 3, 5),
+            (3, 4, 7),
+        ])
+        def test_parametrize(a, b, expected):
+            assert a + b == expected
+    ",
+            package = package,
+            parametrize_function = &get_parametrize_function(package),
+        ),
+    )?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 3
+    All checks passed!
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+fn get_source_code(constructor_body: &str) -> Vec<(String, String)> {
+    vec![
+        (
+            "src/__init__.py".to_string(),
+            "from .calculator import Calculator".to_string(),
+        ),
+        (
+            "src/calculator.py".to_string(),
+            format!(
+                r"
+                class Calculator:
+                    def __init__(self) -> None:
+                        {constructor_body}
+
+                    def add(self, a: int, b: int) -> int:
+                        return a + b
+
+                    def subtract(self, a: int, b: int) -> int:
+                        return a - b
+
+                    def multiply(self, a: int, b: int) -> int:
+                        return a * b
+
+                    def divide(self, a: int, b: int) -> float:
+                        return a / b
+                ",
+            ),
+        ),
+    ]
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_function_scopes(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("print('Calculator initialized')");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator() -> Calculator:
+                    return Calculator()
+                ",
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+
+            def test_subtract(calculator: Calculator) -> None:
+                assert calculator.subtract(1, 2) == -1
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Calculator initialized
+    Calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_module_scopes(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("print('Calculator initialized')");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="module")
+                def calculator() -> Calculator:
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+
+            def test_add_2(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_package_scopes(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("print('Calculator initialized')");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="package")
+                def calculator() -> Calculator:
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+        (
+            "tests/test_calculator_2.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_session_scopes(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("print('Calculator initialized')");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="session")
+                def calculator() -> Calculator:
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+        (
+            "tests/test_calculator_2.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+        (
+            "tests/inner/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 3
+    All checks passed!
+    Calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_mixed_scopes(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="session")
+                def session_calculator() -> Calculator:
+                    print("Session calculator initialized")
+                    return Calculator()
+
+                @fixture
+                def function_calculator() -> Calculator:
+                    print("Function calculator initialized")
+                    return Calculator()
+                "#,
+
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_session_fixture(session_calculator: Calculator) -> None:
+                assert session_calculator.add(1, 2) == 3
+
+            def test_function_fixture(function_calculator: Calculator) -> None:
+                assert function_calculator.add(1, 2) == 3
+
+            def test_both_fixtures(session_calculator: Calculator, function_calculator: Calculator) -> None:
+                assert session_calculator.add(1, 2) == 3
+                assert function_calculator.add(1, 2) == 3
+            ".to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 3
+    All checks passed!
+    Session calculator initialized
+    Function calculator initialized
+    Function calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_fixture_across_files(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="package")
+                def package_calculator() -> Calculator:
+                    print("Package calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_package_fixture(package_calculator: Calculator) -> None:
+                assert package_calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+        (
+            "tests/another_test.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_same_package_fixture(package_calculator: Calculator) -> None:
+                assert package_calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Package calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_named_fixtures(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("print('Named calculator initialized')");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(name="named_calculator")
+                def calculator() -> Calculator:
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_named_fixture(named_calculator: Calculator) -> None:
+                assert named_calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Named calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_nested_package_scopes(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/test_calculator.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator() -> Calculator:
+                    print("Calculator initialized")
+                    return Calculator()
+
+                def test_add(calculator: Calculator) -> None:
+                    assert calculator.add(1, 2) == 3
+                "#,
+            ),
+        ),
+        (
+            "tests/inner/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="package")
+                def package_calculator() -> Calculator:
+                    print("Package calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/inner/sub/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add(package_calculator: Calculator) -> None:
+                assert package_calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Calculator initialized
+    Package calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_independent_fixtures(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator_a() -> Calculator:
+                    print("Calculator A initialized")
+                    return Calculator()
+
+                @fixture
+                def calculator_b() -> Calculator:
+                    print("Calculator B initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_a(calculator_a: Calculator) -> None:
+                assert calculator_a.add(1, 2) == 3
+
+            def test_b(calculator_b: Calculator) -> None:
+                assert calculator_b.multiply(2, 3) == 6
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Calculator A initialized
+    Calculator B initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_multiple_files_independent_fixtures(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="module")
+                def multiply_calculator() -> Calculator:
+                    print("Multiply calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_add.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_add_1(multiply_calculator: Calculator) -> None:
+                assert multiply_calculator.add(1, 2) == 3
+
+            def test_add_2(multiply_calculator: Calculator) -> None:
+                assert multiply_calculator.add(3, 4) == 7
+            "
+            .to_string(),
+        ),
+        (
+            "tests/test_multiply.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_multiply_1(multiply_calculator: Calculator) -> None:
+                assert multiply_calculator.multiply(2, 3) == 6
+
+            def test_multiply_2(multiply_calculator: Calculator) -> None:
+                assert multiply_calculator.multiply(4, 5) == 20
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 4
+    All checks passed!
+    Multiply calculator initialized
+    Multiply calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_basic_error_handling(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def failing_calculator() -> Calculator:
+                    raise RuntimeError("Fixture initialization failed")
+
+                @fixture
+                def working_calculator() -> Calculator:
+                    print("Working calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_working(working_calculator: Calculator) -> None:
+                assert working_calculator.add(1, 2) == 3
+
+            def test_failing(failing_calculator: Calculator) -> None:
+                assert failing_calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: false
+    exit_code: 1
+    ----- stdout -----
+    error[fixture-not-found] in <temp_dir>/tests/test_calculator.py
+     | Fixture failing_calculator not found
+
+    Passed tests: 1
+    Errored tests: 1
+    Working calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_different_scopes_independent(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="session")
+                def session_calculator() -> Calculator:
+                    print("Session calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="package")
+                def package_calculator() -> Calculator:
+                    print("Package calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/inner/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="module")
+                def module_calculator() -> Calculator:
+                    print("Module calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/inner/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_session(session_calculator: Calculator) -> None:
+                assert session_calculator.add(1, 2) == 3
+
+            def test_package(package_calculator: Calculator) -> None:
+                assert package_calculator.subtract(5, 3) == 2
+
+            def test_module(module_calculator: Calculator) -> None:
+                assert module_calculator.multiply(2, 3) == 6
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 3
+    All checks passed!
+    Session calculator initialized
+    Package calculator initialized
+    Module calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_invalid_scope_value(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="invalid_scope")
+                def calculator() -> Calculator:
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_calc(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: false
+    exit_code: 1
+    ----- stdout -----
+    error[invalid-fixture] in <temp_dir>/tests/conftest.py
+     | Invalid fixture scope: invalid_scope
+
+    error[fixture-not-found] in <temp_dir>/tests/test_calculator.py
+     | Fixture calculator not found
+
+    Errored tests: 1
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_invalid_fixture_name(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(name="123invalid")
+                def calculator() -> Calculator:
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_calc(calculator: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: false
+    exit_code: 1
+    ----- stdout -----
+    error[fixture-not-found] in <temp_dir>/tests/test_calculator.py
+     | Fixture calculator not found
+
+    Errored tests: 1
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_multiple_conftest_same_dir(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator_1() -> Calculator:
+                    print("Calculator 1 initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/conftest_more.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator_2() -> Calculator:
+                    print("Calculator 2 initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_calc(calculator_1: Calculator, calculator_2: Calculator) -> None:
+                assert calculator_1.add(1, 2) == 3
+                assert calculator_2.multiply(2, 3) == 6
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: false
+    exit_code: 1
+    ----- stdout -----
+    error[fixture-not-found] in <temp_dir>/tests/test_calculator.py
+     | Fixture calculator_2 not found
+
+    Errored tests: 1
+    Calculator 1 initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_very_deep_directory_structure(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="session")
+                def root_calc() -> Calculator:
+                    print("Root calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/level1/level2/level3/level4/level5/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="package")
+                def deep_calc() -> Calculator:
+                    print("Deep calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/level1/level2/level3/level4/level5/test_deep.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_deep(root_calc: Calculator, deep_calc: Calculator) -> None:
+                assert root_calc.add(1, 2) == 3
+                assert deep_calc.multiply(2, 3) == 6
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Root calculator initialized
+    Deep calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_fixture_in_init_file(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/__init__.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def init_calculator() -> Calculator:
+                    print("Init calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_init_fixture(init_calculator: Calculator) -> None:
+                assert init_calculator.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: false
+    exit_code: 1
+    ----- stdout -----
+    error[fixture-not-found] in <temp_dir>/tests/test_calculator.py
+     | Fixture init_calculator not found
+
+    Errored tests: 1
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_same_fixture_name_different_types(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/math/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def value() -> Calculator:
+                    print("Calculator value initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/math/test_math.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_math_value(value: Calculator) -> None:
+                assert value.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+        (
+            "tests/string/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+
+                @fixture
+                def value() -> str:
+                    print("String value initialized")
+                    return "test"
+                "#,
+            ),
+        ),
+        (
+            "tests/string/test_string.py".to_string(),
+            r#"
+            def test_string_value(value: str) -> None:
+                assert value == "test"
+            "#
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    String value initialized
+    Calculator value initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_fixture_dependencies(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("print('Calculator initialized')");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator() -> Calculator:
+                    print("Calculator fixture initialized")
+                    return Calculator()
+
+                @fixture
+                def calculator_with_value(calculator: Calculator) -> Calculator:
+                    print("Calculator with value fixture initialized")
+                    calculator.add(5, 5)
+                    return calculator
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_calculator_with_value(calculator_with_value: Calculator) -> None:
+                assert calculator_with_value.add(1, 2) == 3
+
+            def test_calculator_dependency(calculator: Calculator, calculator_with_value: Calculator) -> None:
+                assert calculator.add(1, 2) == 3
+                assert calculator_with_value.add(1, 2) == 3
+            ".to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Calculator fixture initialized
+    Calculator initialized
+    Calculator with value fixture initialized
+    Calculator fixture initialized
+    Calculator initialized
+    Calculator with value fixture initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_dependent_fixtures_different_scopes(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="session")
+                def session_calculator() -> Calculator:
+                    print("Session calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="package")
+                def package_calculator(session_calculator: Calculator) -> Calculator:
+                    print("Package calculator initialized")
+                    session_calculator.add(1, 1)
+                    return session_calculator
+                "#,
+            ),
+        ),
+        (
+            "tests/inner/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="module")
+                def module_calculator(package_calculator: Calculator) -> Calculator:
+                    print("Module calculator initialized")
+                    package_calculator.add(2, 2)
+                    return package_calculator
+                "#,
+            ),
+        ),
+        (
+            "tests/inner/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_calculator_chain(module_calculator: Calculator) -> None:
+                assert module_calculator.add(1, 2) == 3
+
+            def test_calculator_chain_2(module_calculator: Calculator) -> None:
+                assert module_calculator.add(3, 4) == 7
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Session calculator initialized
+    Package calculator initialized
+    Module calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_complex_dependency_chain(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def base_calculator() -> Calculator:
+                    print("Base calculator initialized")
+                    return Calculator()
+
+                @fixture
+                def add_calculator(base_calculator: Calculator) -> Calculator:
+                    print("Add calculator initialized")
+                    base_calculator.add(1, 1)
+                    return base_calculator
+
+                @fixture
+                def multiply_calculator(add_calculator: Calculator) -> Calculator:
+                    print("Multiply calculator initialized")
+                    add_calculator.multiply(2, 2)
+                    return add_calculator
+
+                @fixture
+                def final_calculator(multiply_calculator: Calculator, base_calculator: Calculator) -> Calculator:
+                    print("Final calculator initialized")
+                    return multiply_calculator
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_complex_chain(final_calculator: Calculator) -> None:
+                assert final_calculator.add(1, 2) == 3
+                assert final_calculator.multiply(2, 3) == 6
+            ".to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Base calculator initialized
+    Add calculator initialized
+    Multiply calculator initialized
+    Final calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_mixed_scope_dependencies(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="session")
+                def session_base() -> Calculator:
+                    print("Session base initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture(scope="package")
+                def package_calc(session_base: Calculator) -> Calculator:
+                    print("Package calc initialized")
+                    return session_base
+
+                @fixture
+                def function_calc(package_calc: Calculator) -> Calculator:
+                    print("Function calc initialized")
+                    return package_calc
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_mixed_scopes_1(function_calc: Calculator) -> None:
+                assert function_calc.add(1, 2) == 3
+
+            def test_mixed_scopes_2(function_calc: Calculator) -> None:
+                assert function_calc.multiply(2, 3) == 6
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Session base initialized
+    Package calc initialized
+    Function calc initialized
+    Function calc initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_diamond_dependency(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def base_calc() -> Calculator:
+                    print("Base calc initialized")
+                    return Calculator()
+
+                @fixture
+                def left_calc(base_calc: Calculator) -> Calculator:
+                    print("Left calc initialized")
+                    base_calc.add(1, 1)
+                    return base_calc
+
+                @fixture
+                def right_calc(base_calc: Calculator) -> Calculator:
+                    print("Right calc initialized")
+                    base_calc.multiply(2, 2)
+                    return base_calc
+
+                @fixture
+                def final_calc(left_calc: Calculator, right_calc: Calculator) -> Calculator:
+                    print("Final calc initialized")
+                    return left_calc
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_diamond(final_calc: Calculator) -> None:
+                assert final_calc.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Base calc initialized
+    Left calc initialized
+    Right calc initialized
+    Final calc initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_generator_fixture(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def generator_fixture():
+                    yield Calculator()
+                    print("Generator fixture teardown")
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r"
+            from src import Calculator
+
+            def test_generator_fixture(generator_fixture: Calculator) -> None:
+                assert generator_fixture.add(1, 2) == 3
+            "
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Generator fixture teardown
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_fixture_called_for_each_parametrization(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator() -> Calculator:
+                    print("Calculator initialized")
+                    return Calculator()
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            format!(
+                r#"
+                from src import Calculator
+                import {}
+
+                @{}.{}(
+                    "value",
+                    [1, 2, 3],
+                )
+                def test_calculator(calculator: Calculator, value: int) -> None:
+                    assert calculator.add(1, value) == value + 1
+                "#,
+                framework,
+                framework,
+                if framework == "karva" {
+                    "tags.parametrize"
+                } else {
+                    "mark.parametrize"
+                }
+            ),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 3
+    All checks passed!
+    Calculator initialized
+    Calculator initialized
+    Calculator initialized
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_fixture_finalizer_called_after_test(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator() -> Calculator:
+                    print("Calculator initialized")
+                    yield Calculator()
+                    print("Calculator finalizer called")
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r#"
+            from src import Calculator
+
+            def test_calculator(calculator: Calculator) -> None:
+                print("Test function called")
+                assert calculator.add(1, 2) == 3
+            "#
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Calculator initialized
+    Test function called
+    Calculator finalizer called
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[rstest]
+#[case("pytest")]
+#[case("karva")]
+fn test_fixture_finalizer_called_at_correct_time(#[case] framework: &str) -> anyhow::Result<()> {
+    let mut files = get_source_code("pass");
+    files.extend([
+        (
+            "tests/conftest.py".to_string(),
+            format!(
+                r#"
+                from {framework} import fixture
+                from src import Calculator
+
+                @fixture
+                def calculator() -> Calculator:
+                    print("Calculator initialized")
+                    yield Calculator()
+                    print("Calculator finalizer called")
+                "#,
+            ),
+        ),
+        (
+            "tests/test_calculator.py".to_string(),
+            r#"
+            from src import Calculator
+
+            def test_calculator(calculator: Calculator) -> None:
+                print("Test function called")
+                assert calculator.add(1, 2) == 3
+
+            def test_calculator_2(calculator: Calculator) -> None:
+                print("Test function 2 called")
+                assert calculator.add(1, 2) == 3
+            "#
+            .to_string(),
+        ),
+    ]);
+
+    let case = TestCase::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
+
+    allow_duplicates!(assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 2
+    All checks passed!
+    Calculator initialized
+    Calculator initialized
+    Test function called
+    Calculator finalizer called
+    Test function 2 called
+    Calculator finalizer called
+
+    ----- stderr -----
+    "#));
+
+    Ok(())
+}
+
+#[test]
+fn test_stdout_is_captured_and_displayed() -> anyhow::Result<()> {
+    let case = TestCase::with_file(
+        "test_std_out_redirected.py",
+        r"
+        def test_std_out_redirected():
+            print('Hello, world!')
+        ",
+    )?;
+
+    assert_cmd_snapshot!(case.command_with_args(&["-s"]), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
+    Hello, world!
+
+    ----- stderr -----
+    "#);
+
+    Ok(())
+}
+
+#[test]
+fn test_stdout_is_captured_and_displayed_with_args() -> anyhow::Result<()> {
+    let case = TestCase::with_file(
+        "test_std_out_redirected.py",
+        r"
+        def test_std_out_redirected():
+            print('Hello, world!')
+        ",
+    )?;
+
+    assert_cmd_snapshot!(case.command(), @r#"
+    success: true
+    exit_code: 0
+    ----- stdout -----
+    Passed tests: 1
+    All checks passed!
 
     ----- stderr -----
     "#);
