+++ crates/karva_core/src/lib.rs
@@ -7,5 +7,4 @@ pub mod runner;
 pub mod tag;
 pub mod utils;
 
-#[cfg(feature = "testing")]
 pub mod testing;
