--- tests/integration/pyproject.toml
@@ -1,13 +1,0 @@
-[project]
-name = "integration"
-version = "0.1.0"
-description = "Integration tests for the Karva CLI"
-readme = "README.md"
-requires-python = ">=3.9"
-
-[tool.ruff]
-line-length = 120
-
-[tool.ruff.lint]
-select = ["ALL"]
-ignore = ["D", "INP001", "PLR2004", "S", "T201"]
